Move `compontents/` directory to `torchbenchmark/_components`
