+++ test_imports.py
@@ -1,4 +1,4 @@
-import torchbenchmark.models.densenet121
+import torchbenchmark.models
 
 model, example_inputs = torchbenchmark.models.densenet121.Model(
     test="eval", device="cuda", batch_size=1
