+++ torchbenchmark/_components/_impl/tasks/base.py
@@ -7,7 +7,7 @@ import marshal
 import textwrap
 import typing
 
-from components._impl.workers import base
+from torchbenchmark._components._impl.workers import base
 
 
 class TaskBase(abc.ABC):
