+++ torchbenchmark/_components/_impl/workers/base.py
