+++ torchbenchmark/_components/model_analyzer/dcgm/cpu_monitor.py
