+++ torchbenchmark/_components/model_analyzer/dcgm/dcgm_agent.py
