+++ torchbenchmark/_components/model_analyzer/dcgm/dcgm_fields_internal.py
