+++ torchbenchmark/_components/model_analyzer/tb_dcgm_types/cpu_record.py
