+++ torchbenchmark/_components/model_analyzer/tb_dcgm_types/gpu_record.py
