+++ torchbenchmark/_components/test/test_worker.py
@@ -8,10 +8,10 @@ import torch
 from torch.testing._internal.common_utils import TestCase, run_tests
 
 try:
-    from components._impl.workers import base as base_worker
-    from components._impl.workers import in_process_worker
-    from components._impl.workers import subprocess_worker
-    from components._impl.workers import subprocess_rpc
+    from torchbenchmark._components._impl.workers import base as base_worker
+    from torchbenchmark._components._impl.workers import in_process_worker
+    from torchbenchmark._components._impl.workers import subprocess_worker
+    from torchbenchmark._components._impl.workers import subprocess_rpc
 except (ImportError, ModuleNotFoundError):
     print(
         f"""
